Move log to its own modules and add custom support to scroll button
