+++ panel/models/log.py
@@ -1,0 +1,18 @@
+from bokeh.core.properties import List, String
+from bokeh.events import ModelEvent
+
+from .layout import Column
+
+
+class ScrollButtonClick(ModelEvent):
+
+    event_name = 'scroll_button_click'
+
+    def __init__(self, model, data=None):
+        self.data = data
+        super().__init__(model=model)
+
+
+class Log(Column):
+
+    visible_objects = List(String())
